+++ services/pipeline/src/main/java/org/hyperledger/besu/services/pipeline/Pipeline.java
@@ -174,15 +174,16 @@ public class Pipeline<I> {
             if (t instanceof CompletionException
                 || t instanceof CancellationException
                 || t instanceof AsyncOperationException) {
-              LOG.debug("Unhandled exception in pipeline. Aborting.", t);
+              LOG.trace("Unhandled exception in pipeline. Aborting.", t);
             } else {
-              LOG.info("Unexpected exception in pipeline. Aborting.", t);
+              LOG.info("Unexpected exception in pipeline. Aborting.");
+              LOG.debug("Unexpected exception in pipeline. Aborting.", t);
             }
             try {
               abort(t);
             } catch (final Throwable t2) {
               // Seems excessive but exceptions that propagate out of this method won't be logged
-              // because the executor just completes the future exceptionally and we never
+              // because the executor just completes the future exceptionally, and we never
               // need to call get on it which would normally expose the error.
               LOG.error("Failed to abort pipeline after error", t2);
             }